create devices assembly, move voice to devices asm
