+++ Assets/MixedRealityToolkit.Devices.meta
@@ -1,6 +1,7 @@
 fileFormatVersion: 2
-guid: 688be0283e1e8eb4dac92a6dcd0687a1
-TextScriptImporter:
+guid: 8e6d8cdb4c358114589841609d126863
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
