+++ Assets/MixedRealityToolkit.Devices/WindowsVoiceInput/MixedRealityToolkit.Devices.WindowsVoiceInput.asmdef
@@ -1,0 +1,20 @@
+{
+    "name": "Microsoft.MixedReality.Toolkit.Devices.WindowsVoiceInput",
+    "references": [
+        "Microsoft.MixedReality.Toolkit",
+        "Microsoft.MixedReality.Toolkit.Core.Utilities.Async"
+    ],
+    "optionalUnityReferences": [],
+    "includePlatforms": [
+        "Editor",
+        "WSA",
+        "WindowsStandalone32",
+        "WindowsStandalone64"
+    ],
+    "excludePlatforms": [],
+    "allowUnsafeCode": false,
+    "overrideReferences": false,
+    "precompiledReferences": [],
+    "autoReferenced": true,
+    "defineConstraints": []
+}
