+++ Assets/MixedRealityToolkit.Devices/WindowsVoiceInput/MixedRealityToolkit.Devices.WindowsVoiceInput.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 6f2a087266881b2468fca7f267d83fcc
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
