+++ Assets/MixedRealityToolkit.Devices/WindowsVoiceInput/WindowsDictationInputDeviceManager.cs.meta
