+++ Assets/MixedRealityToolkit.Devices/WindowsVoiceInput/WindowsSpeechInputDeviceManager.cs.meta
