--- Assets/MixedRealityToolkit/_Core/Devices/VoiceInput/README.md
@@ -1,3 +1,0 @@
-# Mixed Reality Toolkit - Devices - VoiceInput
-
-Device settings and definitions for gathering speech input (Speech and Dictation), currently only supporting the Windows Platform.
